cmd: refine cmd init
