+++ cmd/root.go
@@ -14,15 +14,9 @@
 package cmd
 
 import (
-	"context"
-	"fmt"
 	"os"
-	"os/signal"
-	"syscall"
 
-	"github.com/pingcap/log"
 	"github.com/spf13/cobra"
-	"go.uber.org/zap"
 )
 
 var rootCmd = &cobra.Command{
@@ -33,26 +27,8 @@ var rootCmd = &cobra.Command{
 
 // Execute runs the root command
 func Execute() {
-	sc := make(chan os.Signal, 1)
-	signal.Notify(sc,
-		syscall.SIGHUP,
-		syscall.SIGINT,
-		syscall.SIGTERM,
-		syscall.SIGQUIT)
-
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-	defaultContext = ctx
-	go func() {
-		sig := <-sc
-		cmdLogMu.Lock()
-		log.Info("got signal to exit", zap.Stringer("signal", sig))
-		cmdLogMu.Unlock()
-		cancel()
-	}()
-
 	if err := rootCmd.Execute(); err != nil {
-		fmt.Println(err)
+		rootCmd.Println(err)
 		os.Exit(1)
 	}
 }
